Improved the log messages.  Finished a preliminary implementation of rolling ball

+++ src/Kernel/RgbPixel.cs
@@ -1,9 +1,9 @@
 ﻿namespace Kernel {
 
     public class RgbPixel {
-        public short Red;
-        public short Green;
-        public short Blue;
+        public byte Red;
+        public byte Green;
+        public byte Blue;
     }
 
 }
